src/main/java/baseball/model/GameModel.java
package baseball.model;

import baseball.utils.Converter;
import camp.nextstep.edu.missionutils.Randoms;

import java.util.ArrayList;
import java.util.HashSet;
import java.util.List;

import static baseball.exception.ExceptionMessage.INVALID_GAMEBALL_SIZE;
import static baseball.exception.ExceptionMessage.NOT_ALLOW_CHARACTER;


public class GameModel {
    private List<Integer> computerNumbers;
    private static final int GAMEBALL_SIZE = 3;
    private static final int STRIKE_COUNT = 3;
    private static final int MIN_GAMEBALL_NUMBER = 1;
    private static final int MAX_GAMEBALL_NUMBER = 9;

    public void initGame() {
        generateComputerNumbers();
    }

    private void generateComputerNumbers() {


        computerNumbers = new ArrayList<>();

        while (computerNumbers.size() < GAMEBALL_SIZE) {
            addComputerNumber();
        }
    }

    private void addComputerNumber() {
        int pickNumber = Randoms.pickNumberInRange(MIN_GAMEBALL_NUMBER, MAX_GAMEBALL_NUMBER);
        if (!computerNumbers.contains(pickNumber)) computerNumbers.add(pickNumber);

    }


    public GameRecord calculateScore(String userNumber) {

        List<Integer> userNumbers = Converter.stringToIntegerArray(userNumber);
        validateUserNumbers(userNumbers);

        int strikeCount = countStrike(computerNumbers, userNumbers);
        int ballCount = countBall(computerNumbers, userNumbers);
        return new GameRecord(strikeCount, ballCount);

    }

    private int countStrike(List<Integer> computerNumbers, List<Integer> userNumbers) {

        int strikeCount = 0;
        for (int i = 0; i < computerNumbers.size(); i++) {
            strikeCount += checkStrike(computerNumbers.get(i), userNumbers.get(i));
        }
        return strikeCount;
    }

    private int checkStrike(int computerNumber, int userNumber) {
        return computerNumber == userNumber ? 1 : 0;
    }

    private int countBall(List<Integer> computerNumbers, List<Integer> userNumbers) {
        int ballNumber = 0;
        for (int i = 0; i < computerNumbers.size(); i++) {
            ballNumber += checkBall(computerNumbers.get(i), i, userNumbers);
        }
        return ballNumber;
    }

    private int checkBall(int computerNumber, int index, List<Integer> userNumbers) {

        int position = userNumbers.indexOf(computerNumber);
        return position != -1 && position != index ? 1 : 0;
    }

    private void validateUserNumbers(List<Integer> userNumbers) {

        if (userNumbers.size() != GAMEBALL_SIZE) throw new IllegalArgumentException(INVALID_GAMEBALL_SIZE);
        if (new HashSet(userNumbers).size() != GAMEBALL_SIZE) throw new IllegalArgumentException(INVALID_GAMEBALL_SIZE);
        for(int i = 0; i < userNumbers.size(); i++){
            checkNumber(userNumbers.get(i));
        }
    }

    private static void checkNumber(int number){
        if( number >= 1 && number <= 9)return;
        throw new IllegalArgumentException(NOT_ALLOW_CHARACTER);
    }


    public boolean isEnd(GameRecord gameRecord) {
        return gameRecord.getStrikeCount() == STRIKE_COUNT;
    }
}
